# docs/openapi.yaml
openapi: 3.0.3

info:
  title: idea-pipeline API
  version: 1.0.0
  description: |
    Toute réponse d'erreur a la forme `{ "error": "…" }`. La liste exhaustive
    des messages vit dans docs/api-design.md, section « Messages » — le serveur
    les reprend mot pour mot.

    Une URL ne correspondant à aucun chemin déclaré ici renvoie
    `404 — Ressource introuvable.`, hors du périmètre de cette spécification.

servers:
  - url: http://localhost:3000
    description: Serveur local — l'API n'est pas hébergée ailleurs.

# Tous les endpoints exigent une session, sauf POST /auth/google qui la crée et
# porte son propre `security: []`.
security:
  - sessionToken: []

paths:
  /ideas:
    get:
      operationId: listIdeas
      summary: Liste des idées
      description: |
        Retourne les idées **du compte appelant**, sans filtrage ni pagination.
        Celles des autres comptes sont invisibles, pas filtrées après coup.

        Triée par le serveur : `updatedAt` décroissant, puis `id` décroissant
        pour départager les égalités. L'ordre est donc déterministe.
      responses:
        '200':
          description: Collection des idées.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Idea'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createIdea
      summary: Crée une nouvelle idée
      description: Crée une idée avec une variation initiale.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdeaRequest'
      responses:
        '201':
          description: Idée créée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `Le champ "text" est obligatoire.` — champ absent, vide, ne contenant que
              des espaces, ou corps non lu faute de `Content-Type: application/json`.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ideas/{id}:
    parameters:
      - $ref: '#/components/parameters/IdeaId'

    delete:
      operationId: deleteIdea
      summary: Supprime une idée
      description: Supprime l'idée identifiée par son id.
      responses:
        '204':
          description: Idée supprimée.
        '404':
          description: |
            `Idée introuvable.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: setIdeaLabel
      summary: Classe une idée à une étape
      description: |
        Rattache l'idée à une étape, ou l'en détache avec `null`.

        L'étape doit appartenir au compte appelant : celle d'un autre répond
        `404`, jamais `403`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetIdeaLabelRequest'
      responses:
        '200':
          description: Idée mise à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `Le champ "labelId" est obligatoire.` — champ absent, ou d'un type autre qu'une chaîne ou `null` ; corps vide inclus.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: |
            Messages possibles :
            - `Idée introuvable.` — l'`id` ne correspond à aucune idée du compte.
            - `Étape introuvable.` — le `labelId` ne correspond à aucune étape du compte.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ideas/{id}/variations:
    parameters:
      - $ref: '#/components/parameters/IdeaId'

    post:
      operationId: addVariation
      summary: Ajoute une variation à une idée
      description: Ajoute une nouvelle variation à l'idée ciblée.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVariationRequest'
      responses:
        '201':
          description: Idée complète avec la nouvelle variation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `Le champ "text" est obligatoire.` — champ absent, vide, ne contenant que
              des espaces, ou corps non lu faute de `Content-Type: application/json`.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: |
            `Idée introuvable.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ideas/{id}/variations/{variationId}:
    parameters:
      - $ref: '#/components/parameters/IdeaId'
      - $ref: '#/components/parameters/VariationId'

    patch:
      operationId: editVariation
      summary: Modifie une variation existante
      description: Modifie le texte d'une variation appartenant à une idée.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditVariationRequest'
      responses:
        '200':
          description: Idée complète mise à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `Le champ "text" est obligatoire.` — champ absent, vide, ne contenant que
              des espaces, ou corps non lu faute de `Content-Type: application/json`.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: |
            Deux causes distinctes, deux messages :
            - `Idée introuvable.` — aucune idée ne porte cet `id`.
            - `Variation introuvable.` — l'idée existe, mais pas la variation ciblée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /labels:
    get:
      operationId: listLabels
      summary: Liste des étapes
      description: |
        Retourne les étapes **du compte appelant**, triées par `position`.

        Un compte sans étape répond `200` avec un tableau vide : c'est un état
        légal, pas une erreur.
      responses:
        '200':
          description: Liste des étapes du compte.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createLabel
      summary: Crée une étape
      description: |
        Ajoute une étape en fin de liste. Le serveur attribue `position` et
        `color` : ni l'une ni l'autre ne se demandent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLabelRequest'
      responses:
        '201':
          description: Étape créée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `Le nom de l'étape est obligatoire.` — nom absent, vide ou uniquement des espaces, corps non lu inclus.
            - `Le nom de l'étape est trop long.` — plus de 32 caractères.
            - `Cette étape existe déjà.` — ce compte porte déjà ce nom, casse ignorée.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: reorderLabels
      summary: Réordonne les étapes
      description: |
        Renumérote les étapes du compte dans l'ordre donné.

        Le corps porte la liste **complète**, et non un déplacement : bouger une
        étape change la position de toutes les autres. Des appels par étape
        s'entrelaceraient en ordres incohérents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderLabelsRequest'
      responses:
        '200':
          description: Étapes renumérotées, dans leur nouvel ordre.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `La liste des étapes est incomplète.` — l'ensemble donné n'est pas exactement celui du compte.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /labels/{id}:
    parameters:
      - $ref: '#/components/parameters/LabelId'

    patch:
      operationId: renameLabel
      summary: Renomme une étape
      description: |
        Seul le nom change. La `position` se modifie par `PATCH /labels`, et la
        `color` ne se choisit pas.

        Aucune idée n'est détachée : elles pointent vers l'identifiant, pas vers
        le nom.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameLabelRequest'
      responses:
        '200':
          description: Étape renommée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '400':
          description: |
            Requête invalide. Messages possibles :
            - `Le nom de l'étape est obligatoire.` — nom absent, vide ou uniquement des espaces.
            - `Le nom de l'étape est trop long.` — plus de 32 caractères.
            - `Cette étape existe déjà.` — une autre étape du compte porte ce nom, casse ignorée.
            - `Champs non autorisés.` — au moins un champ hors du schéma est présent.
            - `Corps de requête JSON invalide.` — JSON malformé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: |
            `Étape introuvable.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteLabel
      summary: Supprime une étape
      description: |
        Les idées qui la portaient deviennent libres. **Aucune idée n'est
        supprimée** : seul le rattachement disparaît.
      responses:
        '204':
          description: Étape supprimée.
        '404':
          description: |
            `Étape introuvable.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/google:
    post:
      operationId: signInWithGoogle
      summary: Échange un code d'autorisation Google contre une session
      description: |
        Le serveur échange lui-même le code auprès de Google, avec son
        `client_secret` — qui ne quitte donc jamais le serveur. Un compte est
        créé au premier passage, retrouvé aux suivants.
      # Le seul endpoint accessible sans session : c'est lui qui la crée.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleSignInRequest'
      responses:
        '201':
          description: La session est ouverte.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: |
            `Code d'autorisation invalide.` — corps invalide, ou code refusé par
            Google.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: |
            `Service d'authentification indisponible.` — Google est injoignable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: |
            `Erreur inattendue côté serveur.`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Le compte associé à la session courante
      description: |
        Permet au client de savoir si son jeton vaut encore, sans avoir à le
        découvrir en échouant sur une requête métier.
      responses:
        '200':
          description: La session est valide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'

  /auth/session:
    delete:
      operationId: signOut
      summary: Déconnexion de cet appareil
      description: |
        Supprime la session portée par le jeton. Les autres appareils du même
        compte restent connectés.
      responses:
        '204':
          description: La session est supprimée.
        '401':
          $ref: '#/components/responses/Unauthenticated'

components:
  responses:
    Unauthenticated:
      description: |
        `Authentification requise.` — jeton absent, invalide, expiré ou révoqué.
        Un seul message pour les quatre causes : les distinguer renseignerait
        un attaquant sur ce qu'il vient de toucher.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        Jeton de session opaque, obtenu par `POST /auth/google`.
        Révocable : supprimer la session le rend immédiatement invalide.

  parameters:
    IdeaId:
      name: id
      in: path
      required: true
      description: Identifiant de l'idée.
      schema:
        type: string
        format: uuid

    VariationId:
      name: variationId
      in: path
      required: true
      description: Identifiant de la variation.
      schema:
        type: string
        format: uuid

    LabelId:
      name: id
      in: path
      required: true
      description: Identifiant de l'étape.
      schema:
        type: string
        format: uuid

  schemas:
    Label:
      type: object
      required:
        - id
        - name
        - color
        - position
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique de l'étape généré par le serveur.
        name:
          type: string
          minLength: 1
          maxLength: 32
          description: |
            Nom de l'étape, écrit par l'utilisateur. Stocké chiffré : la base
            n'en contient jamais le texte lisible.
        color:
          type: integer
          minimum: 1
          maximum: 8
          description: |
            Rang dans une palette de huit, et non une couleur : les valeurs
            vivent dans `src/styles/tokens.css`, seul endroit du dépôt où une
            couleur s'écrit. Attribué par le serveur, jamais choisi.
        position:
          type: integer
          minimum: 1
          description: |
            Rang d'affichage, arrangé par l'utilisateur. Ce n'est pas une
            progression que le produit comprend : toute étape reste atteignable
            depuis toute autre.

    CreateLabelRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32

    RenameLabelRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32

    ReorderLabelsRequest:
      type: object
      additionalProperties: false
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
          description: |
            L'ensemble **complet** des identifiants d'étape du compte, dans
            l'ordre voulu. Un ensemble différent — incomplet, ou portant un
            identifiant inconnu — répond `400`.

    Variation:
      type: object
      required:
        - id
        - text
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique de la variation généré par le serveur.
        text:
          type: string
          description: Texte de la variation.
        createdAt:
          type: string
          format: date-time
          description: Date de création de la variation.

    Idea:
      type: object
      required:
        - id
        - labelId
        - createdAt
        - updatedAt
        - variations
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique de l'idée généré par le serveur.
        labelId:
          type: string
          format: uuid
          nullable: true
          description: |
            L'étape où se trouve l'idée, ou `null` : une idée naît **libre** et
            peut le rester. Toujours présent, jamais absent — le front lit
            `null`, pas `undefined`.
        createdAt:
          type: string
          format: date-time
          description: Date de création de l'idée.
        updatedAt:
          type: string
          format: date-time
          description: Date de dernière modification de l'idée.
        variations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Variation'

    GoogleSignInRequest:
      type: object
      additionalProperties: false
      required: [code, codeVerifier]
      properties:
        code:
          type: string
          minLength: 1
          description: Le code d'autorisation renvoyé par Google au client.
        codeVerifier:
          type: string
          minLength: 1
          description: Le vérificateur PKCE correspondant au défi envoyé à Google.

    Session:
      type: object
      additionalProperties: false
      required: [token, user]
      properties:
        token:
          type: string
          description: |
            Jeton opaque à placer dans `Authorization: Bearer`. Il n'est
            transmis qu'ici : le serveur n'en conserve qu'une empreinte.
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      additionalProperties: false
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: Fournie par Google. Elle peut changer ; `id` non.

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: |
            Message décrivant l'erreur. La liste exhaustive des messages vit
            dans docs/api-design.md, section « Messages » — le serveur les
            reprend mot pour mot.
          example: Idée introuvable.

    CreateIdeaRequest:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          description: Texte initial de l'idée.

    CreateVariationRequest:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          description: Texte de la variation.

    EditVariationRequest:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          description: Nouveau texte de la variation.

    SetIdeaLabelRequest:
      type: object
      additionalProperties: false
      required:
        - labelId
      properties:
        labelId:
          type: string
          format: uuid
          nullable: true
          description: |
            L'étape visée, ou `null` pour détacher l'idée. `null` est une
            valeur comme une autre : il n'y a pas d'endpoint séparé pour
            détacher.